workers/kaxcap-factset/requirements.txt
fds.sdk.Formula==2.2.1
fds.sdk.utils
pandas
requests     
python-dotenv 
